# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Warp API
  version: 0.1.0
  description: |
    Warp's HTTP API. This document is the source of truth: the Go server
    interfaces and the frontend's TypeScript client are both generated from it,
    and a handler that drifts from it fails to compile.

    Two rules from CLAUDE.md are enforced by the contract itself rather than by
    convention:

    * **Every read is context-scoped.** `GET /signals` requires `contextIds`.
      There is no way to ask this API for everything at once.
    * **Raw payloads never leave the system.** A `Signal` carries its subject
      and snippet, never `payload` — that is other people's correspondence, and
      the timeline does not need it.

    **Two things here involve Google OAuth and must not be confused.** Signing in
    (`/auth/google/*`) proves the owner is themselves, asks only for
    `openid email profile`, and produces a short session tied to them being
    present. Connecting a mailbox writes to `accounts` instead, asks for Gmail
    and Calendar scopes, and produces a refresh token that must keep working
    unattended for months. Revoking a session must never stop the overnight sync.

    Scope is Phase 1 (*See*): authentication, contexts, accounts, and the signal
    timeline. Work items, work sessions, and proposed actions arrive with later
    phases.
  contact:
    name: Warp
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development

...
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/auth/google/start:
    get:
      operationId: startGoogleSignIn
      tags: [auth]
      summary: Begin signing in with Google
      description: |
        Redirects to Google. Sets a short-lived, `HttpOnly` state cookie that the
        callback checks, which is what stops a forged callback from signing
        someone in.

        Scopes are `openid email profile` only. This grants no access to mail or
        calendar — connecting those is a separate consent, and lives with
        `accounts`.
      security: []
      parameters:
        - name: returnTo
          in: query
          required: false
          description: |
            Where to land after signing in. A path within this application only:
            the pattern rejects an absolute URL and a protocol-relative `//host`,
            so this cannot become an open redirect.
          schema:
            type: string
            pattern: '^/([^/].*)?$'
            maxLength: 512
            default: /
      responses:
        "302":
          description: Redirect to Google's consent screen.
          headers:
            Location:
              required: true
              schema:
                type: string
            Set-Cookie:
              description: The `HttpOnly` state cookie.
              schema:
                type: string
        default:
          $ref: "#/components/responses/Error"
  /api/v1/auth/google/callback:
    get:
      operationId: completeGoogleSignIn
      tags: [auth]
      summary: Finish signing in with Google
      description: |
        Where Google returns the browser. Verifies `state` against the cookie,
        exchanges `code`, and matches the resulting `sub` claim against
        `auth_providers`. Matching is on `sub`, never on the email address: an
        email can be reassigned at the provider, and matching on it would hand
        the account to whoever inherits it.

        An identity that is not already linked is rejected. There is exactly one
        owner, and sign-in does not create accounts.

        On success, sets the session cookie and redirects to `returnTo`.
      security: []
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          schema:
            type: string
        - name: error
          in: query
          required: false
          description: Present when the owner declined at Google's consent screen.
          schema:
            type: string
      responses:
        "302":
          description: Signed in; redirecting into the application.
          headers:
            Location:
              required: true
              schema:
                type: string
            Set-Cookie:
              description: The session cookie — `HttpOnly`, `Secure`, `SameSite=Lax`.
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthenticated"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: warp_session
      description: |
        An opaque session token. The server stores only its hash, so a database
        dump is not enough to sign in. `HttpOnly`, so a scripting flaw in the
        dashboard cannot read it.
...
  responses:
    Error:
      description: The request failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Unauthenticated:
      description: No valid session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      description: |
        Every error this API returns has this shape, so a client has exactly one
        branch to write.
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/Error"
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable, machine-readable. Safe to switch on.
          example: not_found
        message:
          type: string
          description: |
            Human-readable and safe to display. Never contains payload
            contents, credentials, or anyone's correspondence.
          example: context not found
        details:
          type: object
          additionalProperties: true
          description: Structured detail, when there is any that is safe to return.
...